+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 # Unreleased
 
+No current changes
+
+# v4.10.1
+
 - Consistent `future()` behavior for Interface and Component
 - Support multi-line version arguments